Refactor QR code generation logic and remove unused functions
